♻️ Move application features in specific package
